+++ flutter_lyra/example/lib/methods/initialize_method.dart
@@ -26,6 +26,8 @@ class _InitializeMethodState extends State<InitializeMethod> {
 
   late String publicKey = widget.dataSet.initialPublicKey;
   late String apiServerName = widget.dataSet.initialApiServerName;
+  late String applePayMerchantId = widget.dataSet.initialApplePayMerchantId;
+  late String applePayMerchantName = widget.dataSet.initialApplePayMerchantName;
   var nfcEnabled = false;
   var cardScanningEnabled = false;
 
@@ -35,6 +37,14 @@ class _InitializeMethodState extends State<InitializeMethod> {
 
   void setApiServerName(String newApiServerName) => setState(() {
         apiServerName = newApiServerName;
+      });
+
+  void setApplePayMerchantId(String newApplePayMerchantId) => setState(() {
+        applePayMerchantId = newApplePayMerchantId;
+      });
+
+  void setApplePayMerchantName(String newApplePayMerchantName) => setState(() {
+        applePayMerchantName = newApplePayMerchantName;
       });
 
   void setNfcEnabled({required bool newNfcEnabled}) => setState(() {
@@ -58,6 +68,8 @@ class _InitializeMethodState extends State<InitializeMethod> {
           apiServerName: apiServerName,
           nfcEnabled: nfcEnabled,
           cardScanningEnabled: cardScanningEnabled,
+          applePayMerchantId: applePayMerchantId,
+          applePayMerchantName: applePayMerchantName,
         ),
       );
 
@@ -103,6 +115,18 @@ class _InitializeMethodState extends State<InitializeMethod> {
             value: apiServerName,
             hintText: 'apiServerName',
             setValue: setApiServerName,
+          ),
+          const SizedBox(height: 16),
+          CustomTextField(
+            value: applePayMerchantId,
+            hintText: 'applePayMerchantId',
+            setValue: setApplePayMerchantId,
+          ),
+          const SizedBox(height: 16),
+          CustomTextField(
+            value: applePayMerchantName,
+            hintText: 'applePayMerchantName',
+            setValue: setApplePayMerchantName,
           ),
           const SizedBox(height: 16),
           SwitchListTile(