+++ flutter_lyra/lib/helpers/lyra_options_converter.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/foundation.dart';
 import 'package:flutter_lyra_platform_interface/flutter_lyra_platform_interface.dart';
 
 import '../flutter_lyra.dart';
@@ -13,6 +14,12 @@ class LyraInitializeOptionsConverter {
         apiServerName: lyraInitializeOptions.apiServerName,
         cardScanningEnabled: lyraInitializeOptions.cardScanningEnabled,
         nfcEnabled: lyraInitializeOptions.nfcEnabled,
+        applePayMerchantId: defaultTargetPlatform == TargetPlatform.iOS
+            ? lyraInitializeOptions.applePayMerchantId
+            : null,
+        applePayMerchantName: defaultTargetPlatform == TargetPlatform.iOS
+            ? lyraInitializeOptions.applePayMerchantName
+            : null,
       );
 
   /// convert a [LyraInitializeOptionsInterface] to a [LyraInitializeOptions]
@@ -23,5 +30,8 @@ class LyraInitializeOptionsConverter {
         apiServerName: lyraInitializeOptionInterface.apiServerName,
         cardScanningEnabled: lyraInitializeOptionInterface.cardScanningEnabled,
         nfcEnabled: lyraInitializeOptionInterface.nfcEnabled,
+        applePayMerchantId: lyraInitializeOptionInterface.applePayMerchantId,
+        applePayMerchantName:
+            lyraInitializeOptionInterface.applePayMerchantName,
       );
 }