+++ flutter_lyra/lib/lyra.dart
@@ -20,17 +20,17 @@ class Lyra {
     required this.options,
   });
 
-  /// @template flutter_lyra.lyra.publicKey}
+  /// {@template flutter_lyra.lyra.publicKey}
   /// [Lyra] public key used when initializing the instance
   /// {@endtemplate}
   final String publicKey;
 
-  /// @template flutter_lyra.lyra.options}
+  /// {@template flutter_lyra.lyra.options}
   /// [Lyra] list of options used when initializing the instance
   /// {@endtemplate}
   final LyraInitializeOptions options;
 
-  /// @template flutter_lyra.lyra.getFormTokenVersion}
+  /// {@template flutter_lyra.lyra.getFormTokenVersion}
   /// Get the form token version link to this sdk version
   ///
   /// You need to add it to your paiement request
@@ -41,7 +41,7 @@ class Lyra {
     return formTokenVersion;
   }
 
-  /// @template flutter_lyra.lyra.process}
+  /// {@template flutter_lyra.lyra.process}
   /// Process the paiement
   ///
   /// [timeout] is the duration after which the process will timeout and the
@@ -83,8 +83,8 @@ class LyraManager {
   /// on Android, this function can throw an uncatchable error
   /// Be sure to give the correct publicKey and options
   /// (It won't throw any errors if you are not connected to internet)
-  // TODO: catch the errors on Android
   /// {@endtemplate}
+  // TODO: catch the errors on Android
   Future<Lyra> initialize({
     required String publicKey,
     required LyraInitializeOptions options,