+++ flutter_lyra/test/lyra_test.dart
@@ -1,3 +1,4 @@
+import 'package:flutter/foundation.dart';
 import 'package:flutter_lyra/flutter_lyra.dart';
 import 'package:flutter_lyra/helpers/lyra_options_converter.dart';
 import 'package:flutter_lyra_platform_interface/flutter_lyra_platform_interface.dart';
@@ -5,8 +6,8 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:mocktail/mocktail.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
-mixin PlatformInterfaceMockMixin on Mock implements MockPlatformInterfaceMixin {
-}
+mixin PlatformInterfaceMockMixin on Mock
+    implements MockPlatformInterfaceMixin {}
 
 class MockFlutterLyraPlatform extends Mock
     with PlatformInterfaceMockMixin
@@ -30,12 +31,58 @@ void main() {
         apiServerName: 'apiServerName',
         cardScanningEnabled: true,
         nfcEnabled: false,
+        applePayMerchantId: 'applePayMerchantId',
+        applePayMerchantName: 'applePayMerchantName',
       );
 
       lyra = Lyra(publicKey: publicKey, options: options);
     });
 
     group('initialize', () {
+      test('returns correct lyra instance for iOS', () async {
+        debugDefaultTargetPlatformOverride = TargetPlatform.iOS;
+
+        final lyraKeyInterface = LyraKeyInterface(
+          publicKey: publicKey,
+          options: LyraInitializeOptionsConverter.toInterface(options),
+        );
+
+        registerFallbackValue(
+          lyraKeyInterface.options,
+        );
+        when(
+          () => flutterLyraPlatform.initialize(
+            publicKey: publicKey,
+            options: any(named: 'options'),
+          ),
+        ).thenAnswer(
+          (_) async => lyraKeyInterface,
+        );
+
+        final lyraReceived = await LyraManager().initialize(
+          publicKey: publicKey,
+          options: options,
+        );
+
+        expect(
+          lyraKeyInterface.publicKey,
+          lyraReceived.publicKey,
+        );
+
+        expect(lyraReceived.options.applePayMerchantId, isNotNull);
+
+        expect(lyraReceived.options.applePayMerchantName, isNotNull);
+
+        expect(
+          LyraInitializeOptionsConverter.fromInterface(
+            lyraKeyInterface.options,
+          ),
+          lyraReceived.options,
+        );
+
+        debugDefaultTargetPlatformOverride = null;
+      });
+
       test('returns correct lyra instance', () async {
         final lyraKeyInterface = LyraKeyInterface(
           publicKey: publicKey,
@@ -63,6 +110,10 @@ void main() {
           lyraKeyInterface.publicKey,
           lyraReceived.publicKey,
         );
+
+        expect(lyraReceived.options.applePayMerchantId, isNull);
+
+        expect(lyraReceived.options.applePayMerchantId, isNull);
 
         expect(
           LyraInitializeOptionsConverter.fromInterface(