+++ flutter_lyra_platform_interface/lib/src/lyra.g.dart
@@ -34,6 +34,8 @@ class LyraInitializeOptionsInterface {
     required this.apiServerName,
     this.nfcEnabled,
     this.cardScanningEnabled,
+    this.applePayMerchantId,
+    this.applePayMerchantName,
   });
 
   String apiServerName;
@@ -42,11 +44,17 @@ class LyraInitializeOptionsInterface {
 
   bool? cardScanningEnabled;
 
+  String? applePayMerchantId;
+
+  String? applePayMerchantName;
+
   Object encode() {
     return <Object?>[
       apiServerName,
       nfcEnabled,
       cardScanningEnabled,
+      applePayMerchantId,
+      applePayMerchantName,
     ];
   }
 
@@ -56,6 +64,8 @@ class LyraInitializeOptionsInterface {
       apiServerName: result[0]! as String,
       nfcEnabled: result[1] as bool?,
       cardScanningEnabled: result[2] as bool?,
+      applePayMerchantId: result[3] as String?,
+      applePayMerchantName: result[4] as String?,
     );
   }
 }