+++ flutter_lyra_platform_interface/pigeons/lyra.dart
@@ -26,11 +26,15 @@ class LyraInitializeOptionsInterface {
     required this.apiServerName,
     required this.nfcEnabled,
     required this.cardScanningEnabled,
+    required this.applePayMerchantId,
+    required this.applePayMerchantName,
   });
 
   final String apiServerName;
   final bool? nfcEnabled;
   final bool? cardScanningEnabled;
+  final String? applePayMerchantId;
+  final String? applePayMerchantName;
 }
 
 class LyraKeyInterface {