+++ flutter_lyra_ios/CHANGELOG.md
@@ -1,14 +1,23 @@
-## 0.2.1
+# 0.3.2
 
  - [Android] Upgrade kotlin version
  - [Android] Upgrade gradle
  - [Android] Upgrade material
  - [Android] Upgrade compileSdkVersion
 
-## 0.2.0
+# 0.3.1
 
- - **BREAKING**: remove implementation of the cancelProcess method.
- - **FEAT**: implement timeout property in the process method.
+- **FEAT**: Add apple pay merchant id.
+
+# 0.3.0
+
+- Update a dependency to the latest release.
+- **FEAT**: implements Apple Pay.
+
+# 0.2.0
+
+- **BREAKING**: remove implementation of the cancelProcess method.
+- **FEAT**: implement timeout property in the process method.
 
 # 0.1.2